# doc/api.yaml
openapi: 3.0.3
info:
  title: Title
  description: Title
  version: 1.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/HEIMAHEFA/hotel/1.0.0
  - description: Your localhost
    url: http://localhost:8080
paths:
...
  /rooms:
    get:
      tags:
        - Rooms
      summary: get all rooms
      parameters:
        - name: page
          schema:
            $ref: '#/components/schemas/Page'
          in: query
          required: true
        - name: pageSize
          schema:
            $ref: '#/components/schemas/PageSize'
          in: query
          required: true
        - name: instant
          required: true
          in: query
          schema:
              $ref: '#/components/schemas/Instant'

      responses:
        200:
          description: list of room
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'

    post:
      tags:
        - Rooms
      summary: create room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoom'
      responses:
        200:
          description: the created room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
  /rooms/{id}:
    get:
      tags:
        - Rooms
      summary: get specific room by its id
      parameters:
        - name: id
          schema:
            type: integer
          in: path
          required: true
        - name: instant
          required: true
          in: query
          schema:
            $ref: '#/components/schemas/Instant'
      responses:
        200:
          description: room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
    put:
      tags:
        - Rooms
      summary: update specific room by referring to its id
      parameters:
        - name: id
          schema:
            type: integer
          in: path
          required: true
      requestBody:
        required: true
        description: room with updated details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoom'
      responses:
        200:
          description: updated room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
...
components:
  schemas:
...
    RoomCategoryEnum:
      type: string
      enum:
        - double
        - family
        - suite
      default: double
      example: double
    CreateRoom:
      type: object
      properties:
        description:
          type: string
          example: "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
        categoryName:
          $ref: '#/components/schemas/RoomCategoryEnum'
        roomName:
          type: string
          example: room ABC
    Room:
      type: object
      allOf:
        - $ref: '#/components/schemas/CreateRoom'
      properties:
        id:
          type: integer
          example: 1
        status:
          $ref: '#/components/schemas/RoomStatus'
        price:
          $ref: '#/components/schemas/Price'
        bookingCount:
          type: integer
          example: 0
    RoomStatus:
      type: string
      enum:
        - AVAILABLE
        - TAKEN
      default: AVAILABLE
    Price:
      type: number
      format: double
...
    Page:
      type: number
      minimum: 1
    PageSize:
      type: number
      minimum: 1
      maximum: 500